Reduce code duplication and test count
